+++ src/main/java/stringcalculator/domain/Separator.java
@@ -1,4 +1,4 @@
-package stringcalculator;
+package stringcalculator.domain;
 
 import java.util.Arrays;
 import java.util.List;
