+++ src/test/java/stringcalculator/NumberTest.java
@@ -5,6 +5,7 @@ import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import java.util.Arrays;
 import java.util.List;
 import org.junit.jupiter.api.Test;
+import stringcalculator.domain.Number;
 
 class NumberTest {
