+++ src/test/java/stringcalculator/SeparatorTest.java
@@ -6,6 +6,7 @@ import java.util.Arrays;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import stringcalculator.domain.Separator;
 
 class SeparatorTest {
